requirements.txt
flask==1.1.1
Flask-Menu==0.7.1
flask-paginate==0.5.5
fastparquet==0.3.3
gin-config==0.3.0
sqlalchemy==1.3.8
pandas==0.25.1
absl-py==0.9.0
psycopg2==2.8.3

